OF-810 RFC 6121 Routing Compliance Fix
This fix fixes the routing, which is described in RFC 6121 8.5.3.2.1. Message

Misbehavior before: Every message addressed to a full JID, whose resource was not available was routed to the bare JID (as fallback). 8.5.3.2.1 specifies, that only chat messages should be routed to bare JID and other message types should either be ignored or bounced with an error.

Behavior now: non-chat messages are delegated to the offline message strategy, which decides, if the message should be ignored or bounced with an error, depending on the settings in offline-messages.jsp.