Update StatsAction.java
The Server traffic statstics in quick stats are incorrect and inconsistent compared with All Reports page.
The problem is in the calculation of Low and High Values.
The StatsViewer.getMin() and getMax() will return two values for high and low for server_bytes_in and server_bytes_out respectively.
The StatsAction.getLowAndHigh(String key,  long[] timePeriod) does not take that into account and always precesses the first element in the array. Modified so that it sums the values of all the elements of the array.