OF-910 - Error stanzas to MUC room should remove occupants
Currently, only stanzas containing errors of type cancel will cause the
occupant to be removed, however this doesn't appear to catch and remove many
occupants in practise.

This patch therefore removes occupants from the service on any message or
presence error stanza, irrespective of the error type. IQ stanzas do not
trigger the same response, as these may be due to lack of client support
rather than a vanished occupant.