User not kicked from room after banning.
After banning, the user is not kicked from the room. The user is added
to the ban list so they can't re-join the room or post messages, but
they are not kicked out of the room so the user can still receive messages
sent to the room.

The problem is an additional check for isMembersOnly() was only allowing users
to be kicked from members only chat rooms. Whenever a user is marked as an
outcast they should be kicked from the room regardless of if the room is a
members only room or not.

Looks like this problem was introduced in commit
822abb014741b82d9248c3f747bea64779546999