Unallow addMember request with empty nicknames for rooms with reserved nicknames.
If there is an addMember request, say from a room admin, with an empty nickname,
then an empty string is assigned as the nickname. For rooms with reserved nicknames,
this will create problems for the following scenario - A room admin grants membership
to a user JID, but with an empty nickname. The user then requests to join the room
with a non-empty nickname, but the room does not allow that since the reserved nickname
does not match with the provided one.