OF-1141 Fix default configuration of PEP
The PEP Service configures leaf nodes to store -1 items by default; however
this value is not corrected by the usual post-configuration checks to 1, and
this in turn means that when gathering the items to return, the single
item is removed.

This patch corrects the issue by setting the maxPublishedItems in the default
configuration to 1 instead.