PrivacyLists on first run There are 2 instances of PrivacyListProvider (one in PrivacyListManager and the other in IQPrivacyHandler)
The issue was : both instances have separate privacyListCount variables, which are filled during PrivacyListProvider constructor (== server startup).
Thus, if at server startup the privacyListCount is 0, and one of the instances adds a PrivacyList, the other will never be informed and discard all changes and reads.

Fix : turn PrivacyListProvider into a singleton.