OF-1229 Safeguard JDBC.executeBatch() calls on empty batch
OF-955 updated the HSQL library and it introduced a JDBC requirement that
executeBatch() calls actually have batched items to do.  It appears there
is no API available to check for if batched statements have been added, so
we hack in a simple sentinal to safegaurd calls to `executeBatch`.  Thankfully,
only PubSubPersistenceManager.java uses this API.  Otherwise, an option is to
set `allow_empty_batch` on the connection parameters.