OF-477: Improve default XMPP domain name and FQDN
This commit intends to have better default values for two properties of Openfire:
- the XMPP domain name (eg: igniterealtime.org)
- the FQDN of the server (eg: xmpp.igniterealtime.org)

Older code did not always make a clear distinction (ambiguously referring to
'server name'). Having a proper FQDN has become more important, as newer SASL
implementations depend on the client using the same FQDN as what is configured
in the server).

This commit removes support for changing the XMPP domain name at run time (although
you can obviously still change the corresponding property). Instead, the FQDN can
now be changes (although will require a restart).

Additionally, the XMPPServerInfo implementation got a bit of a refactoring: values
that are not expected to be modified (or considered unmodifiable from a functional
perspective), are no longer passed around, but are rather hardcoded in the
implementation. Support for deprecated ServerPort was removed.

Finally, most access to the properties that hold the XMPP domain name and FQDN
are now encapsulated by XMPPServerInfoImpl.