Add core security labelling support
This design presumes the existence of an "Access Control Decision Function", which can take abstract clearances (as a String) and the XEP-0258 security label construct and evaluate them.

The function - actually an object - must be able to obtain a clearance string given a Jid, perform a check based on a clearance - optionally rewriting the label suitable for a particular Jid - and finally check the label for validity.

An abstract ACDF is provided which will handle the getClearance interface.

A deployment may only have a single AccessControlDecisionFunction.