OF-546: Update HSQLDB dependency to match included library.
The HSQLDB library that ships with Openfire (for Ant) is newer than the
dependency that's defined in Maven. This introduces a compatibility issue
at runtime:

> org.jivesoftware.database.SchemaManager - Unexpected token: FILES in statement [ SET FILES]

The difference is likely an oversight, caused by us taking a very long time
to merge the Maven PR. At the time it was created, the dependency was still okay.