OF-546: Allow plugin name override.
By default, the Maven structure uses the Maven artifactId value as the name for a plugin.
Maven artifactIds typically include tokens that will cause problems, such as dashes and dots.
The plugin name is used in various places, that all have limitations: filenames (the jar file
itself, the i18n files), package names (for the java files generated from JSPs), etc.

This commit allows the name of the plugin to be overridden using a maven property 'plugin.name'
and includes instructions on what the best format is.