OF-1324: Queue that holds outbound data pending s2s session should be bounded.
This commit adds an upper-limit to the amount of stanzas that can be queued in an outgoing
session promise. Any data that won't fit in the queue is rejected immediately. Note that
this can lead to a situation where later stanzas are rejected before stanzas that or sent
earlier (but got in the queue).