OF-546: Add metadata filter
This commit allows a plugin.xml to use a reference to data defined in the pom.xml. This reduces duplicate data.

For instance, in plugin.xml, one could now define the version number by referring to the version number of the
maven artifact:

  <version>${project.version}</version>

Note that plugin.xml files that are processed using the old Apache Ant based build, will not be processed correctly.