OF-1339: Prepare websocket plugin for merge into core.
Openfire 4.2.0 will have the functionality that's now in the websocket plugin. As a result,
this plugin should not be used in Openfire 4.2.0. To facilitate this, we've defined a new concept
of 'priorToServerVersion' in OF-1353.

This commit defines that the plugin should only be used in Openfire instances prior to version 4.2.0.