Revert "OF-1542: Apply filtering to the user summary page"
This reverts (most of) commit e1ee7a3c75b8110f921ae23b038d32cdd086d757.

The changes in the original commit caused all users to be obtained from
the user provider when looking at the user-summary page. For instances
with many users, this requires to much resources and takes to long to
complete.