OF-1596: Allow connections to properly close
The notification will trigger some shutdown procedures that, amongst other things,
check what type of session (eg: anonymous) is being closed. This check depends on the
session still being available.

For that reason, it's important to first notify the listeners, and then close the
session - not the other way around.

This fixes a very visible bug where MUC users would remain in the MUC room long after
their session was closed. Effectively, the bug prevents the MUC room from getting a
presence update to notify it that the user logged off.