OF-1701: The same component can connect to more than one cluster node.
An external component can connect more than once to the same cluster node. When that is true, then it stands to reason
that an external component can also be connected to more than one cluster node.

The old code only tracked one cluster node for each connected external component. It is modified to track each cluster
node to which a component is connected.

Additionally, this commit replaces the byte-array representation of a node ID with instances of NodeID (which mimics
the behavior of RoutingTableImpl.