OF-1706: Server instances should have a unique Node ID.
In a cluster, servers are likely to briefly run in a non-clustered mode. In that mode, they use a default node ID, that can be used to reference data.

This ID changes when the server joins a cluster. The references should be updated, but this is error-prone (see OF-1705).

It is useful for the local node to continue to be able to find the references that it created. This cannot reliably be done when every cluster node uses the same default value.

To reduce the risk of problems, every cluster node should get a unique default node ID.