Refactor: Component session setup/teardown responsibilities
This commit aims to reduce complexity of the code, without introducing functional changes.

Existing code was creating and tearing down component sessions in various places, which made the process as a whole not very transparant, but error prone. This commit changes logic in such a way, that a component that creates a Component Session, also takes responsibility for making sure it's cleaned up.