OF-1700: Cluster teardown should be reflected in SessionManager state.
This fixes an issue that is similar to the one described in OF-1700. When a cluster node leaves the cluster, the state as managed in org.jivesoftware.openfire.SessionManager should be updated to reflect this.

As far as I can tell, there have not been reports of issues related to the invalid state of SessionManager (unlike OF-1700, that caused External Components to remain visible while they were not). Nevertheless, maintaining proper state should help prevent issues.