OF-1705: Removed redundant 'cache restore' from RoutingTableImpl
The code removed by this commit 'solved' OF-1705 by repeating the method invocations that eventually lead to adding entries to the cache, updating the NodeIDs in the processes.
OF-1705 is much more precise: instead of going through the process of completely re-computing the to-be-re-added values, it simply modifies the existing values. This is more
concise, and hopefully more performant.