OF-1700: Cluster teardown should be reflected in RoutingTableImpl's componentsCache state.
This fixes an issue that is similar to the one described in OF-1700. When a cluster node leaves the cluster, the state as managed in org.jivesoftware.openfire.spi.RoutingTableImpl should be updated to reflect this.

Unlike classes that got similar changes in other commits, RoutingTableImpl already cleaned up state when one node left the cluster (but not when
the local node left the cluster). This commit replaces that code with the more concise constructs now used in other places.

As far as I can tell, there have not been reports of issues related to the invalid state of RoutingTable (unlike OF-1700, that caused External Components to remain visible while they were not). Nevertheless, maintaining proper state should help prevent issues.