OF-1700: Cluster teardown should be reflected in RoutingTableImpl's usersCache, anonyousUsersCache, usersSessions state.
This fixes an issue that is similar to the one described in OF-1700. When a cluster node leaves the cluster, the state as managed in org.jivesoftware.openfire.spi.RoutingTableImpl should be updated to reflect this.

Unlike classes that got similar changes in other commits, RoutingTableImpl already cleaned up state when one node left the cluster (but not when
the local node left the cluster). The principle isn't easily replaced by a similar construct as the one used for OF-1700. This commit does apply
the fix to the event handler user when the local node leaves the cluster.

As far as I can tell, there have not been reports of issues related to the invalid state of RoutingTable (unlike OF-1700, that caused External Components to remain visible while they were not). Nevertheless, maintaining proper state should help prevent issues.