OF-1729: Clean up user-specific cache entries upon user deletion.
When a user is deleted, cached data for that user should be deleted. This was generally already the case, but some issues were found with regards to "offline presence", "last activity" and "entity capabilities" which is fixed in this commit.

Additionally, the EntityCapabilitiesManager has been turned into a proper Module (which allows for teardown, needed to properly deregister as a user event listener).