OF-1707: Should not send disco#info to already available components.
Openfire allows more than one instance of a component to operate behind a JID. This is mainly used for load-distribution.

When a new Component connects, a service discovery request is sent to it. Its response is used to trigger event listeners.

If the subdomain of the component is already active (if another than the first component connects for a domain), then doing service discovery is not needed. The response triggers unwanted events in registered ComponentEventListener instances.