OF-1799: Allow JID caches to be configurable
The JID class uses three caches, in which the result of stringprepping is stored. The capacity of these caches is hardcoded.

For high-volume instances of Openfire, the hardcoded values might not suffice. It should be possible to reconfigure the caches. Additionally, it would be nice if the caches are displayed in the cache summary admin console page.

This commit pulls in a new version of the Tinder dependency, that is as of yet unreleased (I'm using a SNAPSHOT version here). Before we release Openfire with this change, this dependency must be replaced with a proper, non-SNAPSHOT release.