OF-1829: Prevent NPE when removing from a non-cached collection.
When multi-valued caches are used, removing an element from a collection that is not (or no longer) in the cache should not throw a null pointer exception.

I suspect that in clustered environments, where caches are managed by all cluster nodes, there's a chance of the same removal to be executed by each node on the
same cache, which could trigger an issue like this.