Make in-memory pubsub implementation compatible with clustering.
This commit primarily adds Serialization to various pubsub implementations, and utilizes that to make state of the new in-memory pubsub provider shareable in a cluster.

To facilite this, a 'unique identifier' class was added to each of the three primary pubsub entities: service, node and item.

The in-memory implementation makes use of multi-valued caches. To properly process thier modification, the CacheUtil class that was originally implemented as part of OF-1700 is added to this commit.