OF-1990: pubsub: split cache from persistence
Historically, the database-based Pubsub persistence provider relies on heavy caching. This mechanism is complex and therefor error prone. It's also not needed for light usage.

This commit splits off the caching implementation from the database-based pubsub persistence provider, into a provider that can be used as a proxy to other providers (to add a layer of caching to the delegated provider). This reduces the code complexity of the database-based implementation, and, by isolating caching code in a dedicated class, allows for easier maintenance of that code too.

This brings these additional changes:
- the 'retry' functionality (where failed database writes are retried a configurable amount of time) has been removed.
- 'flush' functionality, that only exists to the state of persistent storage in line with caches, has been moved out of the interface definition (it now lives solely in the CachingPubsubPersistenceProvider)