Added support for flatting nested groups.
Nested / complex / hierarchic groups are flattened. This means: if
group A is member of group B, the members of group A will also be
members of group B.

Cyclic paths are detected and followed only once.

The features is disable by default.
Set property "ldap.flattenNestedGroups" to "true" to enable it.