OF-353: LDAP Remove support for 'enclosed' DNs.
As a resolution for JM-1327, the concept of 'enclosing' DNs with double-quote (") characters was introduced.

The previous commit adds more elaborate escaping of reserved characters, which makes this 'enclosing' redundant. As the code now uses javax.naming.* types, instead of Strings, to represent DNs, it'll be more challenging to restorte the 'enclosing' functionality.

To reduce code complexity, the 'enclosing' functionality (that no longer serves any purpose) is removed in this commit.