OF-353: Revert from LdapName to String for adminDN
Earlier commits introduces the usage of LdapName to represent DNs. This had an unexpected side effect: the LDAP admin DN now strictly requires the full DN.

In the past, the userprincipalname format (user.name@domain.com) could be used, but not anymore. This will likely break some installs.

This commit reverts back to String to represent the adminDN.