OF-353: Improve escaping for JNDI
As LDAP is being accessed through JNDI, data needs to be handled in a way that is aware of JNDI specifics.

For input parameters, this was already taken into account. A similar handling however, failed for processing search results. This commit adds that.

Additionally, the implementation responsible for this is no longer custom code. Instead, the LdapName class (provided by Java) is now used for this purpose.