OF-1956: Introduce a configurable pause when restarting connection listeners.
To combat race conditions when restarting a (socket) listener, this commit introduces a configurable pause that occurs when restarting a listener (after 'stop', before 'start').

As the race condition is more likely to affect the older s2s code, a non-zero default has been configured for that listener. Client listeners have a default of zero.