OF-1968: Add vCard (avatar) support for MUC rooms.
This largely follows the specification as defined by ProcessOne: https://docs.ejabberd.im/tutorials/muc-vcard, with an addition to make things work with Converse.

The change herein stores a room vCard into the same database table that's used for regular users (but using the JID for the room, as opposed to the username that's used for users).

A downside of this approach is that setting avatars for MUC rooms likely does not work in AD/LDAP setups (or other setups where the vCardProvider is read-only).

This commit lacks an admin console change (it would be nice to also be able to specify a vCard or avatar for a room through the admin console).