OF-158: Make room default settings use configurable service defaults.
MUC services have configurable default settings to be applied to new rooms. These defaults are already used when an end-user creates a room (through XMPP).

This commit takes applies the same settings to the room creation form that is in the Admin console, with one exemption: rooms created in the admin console are always persistent.

Note that it's possible to configure Openfire with more than one MUC service. If that's the case, selecting the MUC service in which the room is to be created is part of the room creation form. As at that point, the to-be-used service is still unknown, the (other) default values that are used on the same form are based on the first available service. That seems to be an acceptable trade-off, given the limited amount of installs that have more than one MUC service.