Add aioxmpp-based tests to Travis
This commit adds a script, similar to the one for the Smack-based integration tests, that downloads and excutes the tests that are part of aioxmpp (https://github.com/horazont/aioxmpp)

aioxmpp requires anonymous authentication to be enabled. This commit achieves that by adding a corresponding autosetup configuration option to the demoboot config file.