OF-872: Properly process IQ and Message without 'to' attribute.
RFC 6120 § 10.3. dictates that IQ and Message stanzas that have no 'to' attribute value should be processed as if the 'to' attribute value equals that of the bare JID of the sender.

This commit parses all inbound IQ and Message stanzas, setting the 'to' attribute value for stanzas that have none to the bare JID that matches the 'from' attribute value. A new system property `xmpp.server.rewrite.replace-missing-to` has been introduced that can be used to disable this functionality.