OF-2015: Clean up the donor immediately after session resumption.
When Stream Management is used to resume an old session, the session that was used to resume the old session should be cleaned up immediately. Before this fix, the session lingers in a state of being 'detached'. Not only is there no reason for it to linger (it can't be resumed), marking it as 'detached' is semantically wrong: this session never reached a state where it makes sense to resume it later. The changes in this commit clean up the donor session immediately after resuption.