OF-2019: MUC archive worker need not start until work is produced
The Archive implementation (which is used to persist chat room history) polls quite aggresively. This has proven to be a problem in an environment that has many MUC services, as many threads performing these polls require quite some CPU resources.

When logging does not take place (eg: when it has been disabled) all of this is quite wasteful. This commit introduces a latch that, if used, delays the pollig until after the first message to-be-logged has been produced. This intends to prevent polling to take place on services where not a single message is being logged.