OF-1990 / OF-1989: Ensure that pubsub items can be overwritten by using the same ID.
XEP-0060 (section 7.1.2) states that if the publisher previously published an item with the same ItemID, successfully processing the request means that the service MUST overwrite the old item with the new item.

Prior to the changes in OF-1990, Openfire probably did this implicitly, through its use of caches. After these changes, duplicate key SQL errors occur when an item with a pre-existing ItemID is being saved (conditionally - when the Cached provider is used, the error seems to be gobbled up).

A similar error is present in the InMemory provider, which will simply add (not replace) the new Item to the node. As test cases use the default max_items value of 1, this has gone unnoticed during testing.