OF-1985: Seperate pubsub service creation that requires service to exist.
When creating a pubsub/pep service, some initialization requires the service to exist. This leads to a chicken/egg problem.

This commit splits off the code that requires the service to have been constructed from the constructor, and calls that initialization only after the newly created service has been registered.