OF-2046: Ignore comments in sidebar-admin.xml
Under certain conditions (that I have been unable to reproduce) adding comments in the admin-sidebar.xml appears to influence the rendering of the admin console.

It has been reported that reconfiguring the SAX reader to ignore comments prevents the issue from occurring.

Although I've not been able to reproduce the issue, ignoring comments is sufficiently benign that it doesn't hurt to add that to the implementation.