OF-2061 & OF-2062: Update default configuration for existing PEP services
Existing PEP services each have a database record of what the default configuration of their nodes is. For PEP, it's very desirable to make two changes:
- ensure that payload of items is persisted
- increase the maximum allowable payload size

This commit adds an update script that modifies the database-stored default configuration for leaf nodes of existing PEP services (identified as services for which the ID matches a JID - the script detects this by looking for an @ character).

Similarly, the configurations of all existing nodes is updated accordingly, provided that the configuration does not appear to have been modified from the default (which is detected by comparing the maximum allowable payload size again).