OF-2080: Prevent caching null item
When an item is being requested that does not exist, the existing code for the CachingPubsubPersistenceProvider attempts to cache null, which is not allowed.

This modification simply skips caching the value. We could consider adding an Optional (to reflect the negative lookup), but that needs additional checking. I'm not sure if it'd work with the async persistence of content that's implemented in this class.