Rewire pluggable persistence provider for Pubsub
The pluggable persistence provider for Pubsub that got introduced in OF-1988 has a lifecycle: it's gets started and stopped. It is crucial that users of the persistence providers do not interact with it unless it is running properly.

This commit explicitly ties the lifecycle of persistence providers (through its manager) to the PubsubModule. This should help ensure that code won't be able to use a persistence provider unless other pubsub code has been initialized, or up until other pubsub state has been teared down.

The changing in this commit work up towards a fix for OF-2102, which aims to fix the shutdown logic of Pubsub services.

The this commit changes the signature of PubsubPersistenceProviderManager. As this code was not in a released version of Openfire yet, I didn't bother to deprecate methods that I've replaced/removed.