OF-2092: Fix PEP memory leak
The original code registered a Entity Capability listener for every PEP Service. That implies that every PEPService instance for all users needs to remain in memory at all times, or at least while any of the owner or subscribers is online. If that isn't the case, then changes in notification filters (expressed through CAPS) might be missed, causing notifications to (not) be sent out when (in)appropriate.

Also in the original code is an issue where the Entity Capability listener is evicted from its cache (which is time-limited), only to be replaced by another instance. As the original instance could still be active elsewhere, this newly added instance would add to the total memory usage, effectively causing a memory leak of sorts.

This commit replaces the 'one-listener-for-each-pepservice' strategy with an implementation that registers one listener that's shared by all PEP services. This removes the need of complex lifecycle handling, and offers better opportunities to work with use cases that involve 'offline' PEP service owners.