OF-2112: Explicity discard S2S-outbound stanzas without a from attribute
Stanzas that originate from clients _always_ have a 'from' attribute (as that attribute value is set
overwritten by Openfire upon receiving the stanza, to prevent abuse where a user tries to impersonate
someone else). That means that, if we're processing a stanza without a 'from' attribute, that the
stanza is very likely to originate from Openfire's code. If we have code that generates a stanza
without a 'from' address but addressed to a remote domain, this simply is a bug that we should very
verbosely warn about.