Improve SubscriptionTask implementation
The existing implementation lazily-loaded a subscription instance. Construction of such an instance is very cheap. Constructed instances are unlikely to be reused often by the SubscriptionTask that created them.

To reduce complexity (and improve maintainability) of the code, the lazy-loading of the subscription instance is being removed.

This should have not functional change. There is an API change that I deem neglible: a transient, package-protected field (that is not accessed outside of the class) is being removed.