OF-2156: Account for async processing of BOSH requests
This is a major refactoring of the BOSH implementation, that primarily aims to address an issue where Openfire does not properly handle BOSH requests that are processed out-of-order (of their request ID / RID) values. These ordering issues can be expected to happen, due to the nature of the mechanism used (HTTP request), as well as because of the asynchronous Jetty API that's used.

Additional improvements have been included (replacing a JiveGlobal with a SystemProperty, improving documentation, fixing most warnings generated by Intellij).