OF-2161: Ensure that SSID for MUC PMs is the MUC JID
When a private message that's exchanged in a MUC room is processed, it receives a stable and unique stanza ID. As per specification, Openfire adds a 'by' attribute. However, for private messages, this value is a full JID (representing the nickname of the user that sent the message), instead of a bare JID (representing the room in which the message was exchanged). This is fixed by ensuring that the JID used is a bare JID.