OF-2163: Prevent unexpected stanza modifications to leak information
Delivering a stanza to a MUC room occupant changes the addressing of the stanza (the room JID is replaced by the real JID). It is often undesirable that this change is applied to the original stanza (that can be processed further), as such changes can leak the privacy-sensitive real address of the recipient.

This commit creates defensive copies of the stanza that is being sent, to prevent leaking the real address of the recipient.