OF-2272: Making MUCRole clusterable: replace user with JID
Instead of keeping a reference to the complex MUCUser instance, the MUCRole implementation now keeps a reference to its JID, using that to look up the desired instance when needed. This makes it easier to share a MUCRole instance in a clustered cache.

To allow the MUCUser to be looked up by JID, additional API was added to MultiUserChatImpl. This needs some cleaning up (notably, the API is not defined in the interface, which is worked around by using a cast, and thread-safety has not been taken into account at all).