OF-2245: Replace volatile boolean with AtomicBoolean
It is hard to reason over volatile booleans (or any member, for that matter). With volatile booleans, a compare-and-set strictly speaking isn't atomic, but there's the 'happens-before' guarantee for volatile members, that might make things safe again. In all, I'm not sure if the volatile member tracking the 'isClosed' state of a HTTP session is sufficiently thread safe. Replacing it with AtomicBoolean will ensure it is, and will at the very least remove ambiguity.