OF-2245: Remove redundant code
In the last commit (86fb7d63a31717b1d3adfd1d5c58eb33a87bcbd3) the state of a HttpSession was made to be equal to the state as maintained by its (Virtual)Connection. With this change, the behavior around `close()` and `isClosed()` is equal to that of the parent class of `HttpSession`: `ClientSession`. There no longer is a need to override these methods in `HttpSession`.