OF-2251: Make BOSH threadpools more configurable
This adds properties for minimum amount of, as well as the TTL for surplus threads in the threadpool used for BOSH network IO.

An additional property is added for the minumum amount of threads in the pool used for BOSH packet processing (which up to now was hard-coded to use 25% of the max).

When switching from JiveGlobals to SystemProperty, there's no need to explicitly call 'migrateProperty' anymore.