OF-2259: Re-use XML parser when loading messages from database.
When profiling Openfire that was starting up from a database that contained a lot of MUC rooms with a lot of messages (ofMucConversationLog), a hotspot was identified in the code that parses the database content into XMPP messages. The parser that is used was recreated for every message, which added considerable overhead.

This commit ensures that the parser that parses database content into messages (as used during the initial load at boot-time) is re-used.