OF-2259: Allow pre-load of MUC rooms to be configurable
Up until this commit, Openfire would always, when it started, load in memory all MUC rooms for which activity was recorded in the last X days. On instances with large amounts of rooms, this leads to a significant delay in boot time.

This commit makes this pre-loading behavior configurable. The amount of days can be configured (independently from the 'cleanup' duration, which it was previously tied to) but the behavior can also be disabled completely.