fixes OF-2266: Do not allow clean-ups interfere with test execution
As cleanup starts pretty much at the same time as the test, it's possible that during the test, an earlier attempt gets cleaned up (causing the test to fail). This has been observed on CI with Github Actions.

By using a mock of the TaskEngine, the cleanup task never gets executed, which prevents the issue.