Add warm-up connection to connectivity tests
For reasons beyond me, one connection (out of many) for the BOSH connectivity test seems to fail - but only on the infrastructure for Github Actions.

From the output, it appears to be the _first_ connection that fails. I suspect that this is triggered by a weird scheduling/timing/eventing thingy that is specific to that environment.

This commit adds an additional connection attempt, for which the outcome will be ignored. Hopefully by 'warming up' the system under test, we get more stable results.