Disable unstable Smack Integration Test
Smack uses an API for two tests (SoftwareInfoIntegrationTest and EntityCapsTest) that itself defines to be 'unstable'). In Openfire CI, we see these tests occasionally fail. It stands to reason that that's because of the Smack-based unstable solution.

This commit disables both tests (one already was disabled), so that they don't break our build.