OF-2220: Ensure that MUC entity changes are explicity re-added to Cache
When an entity that's in a Cache is being modified, these modifications will only become visible to other cluster nodes after the entity has been updated in the cache. This commit ensures that this happens for MUC-related entities.

After all Local- and Remote MUC entities have been replaced by a singular entity (OF-2272), and those entities were shared across the cluster, the need to work with (String-based) references to those entities (eg: room name, instead of MUCRoom) has greatly diminished. This commit contains changes that pass around MUCRoom entities instead of room names (which are then used to look up the corresponding MUCRoom instance in various places). This reduces the amount of code locations where MUCRooms are being obtain from (and put in) a Cache.

This commit ensures that all code that obtain a MUCRoom from a Cache takes responsibility for 'putting back' the instance in the Cache if modification are (likely to have been) made.

To simplify the implementation, a MUCRoom-based, cluster-wide lock is obtained when there is any interaction with a MUC room. This is largely, but not exclusively, done when a stanza arrives to be processed by the room. Before releasing the lock, the Cache entry is updated/synced (as stated above).

There are a few exceptions, where a Lock is _not_ obtained, for example when code iterates over all rooms in the server, while performing read-only operations on them. The cost of locking all rooms is deemed to high for the benefit of transaction safety that they bring, in these cases.