OF-2272: Fix serialization of MUCRoomHistory
Serialization of MUCRoomHistory fails when a field of the serialized data is the MUCRoom, that in turn has the MUCRoomHistory instance as part of it's serialized data (a loop will occur when deserializing this tangle).

This code modifies MUCRoomHistory to serialize only a reference to the room, instead the room itself, which prevents this loop.