OF-2220: Ensure that MUCUser is updated in cache after changes
When an instance that 'lives' in a clustered cache is being modified, the modified instance needs to be put back in the cache explicitly. If that's not done, the changes will not be visible when the object is re-obtained from the cache (eg: by other cluster nodes).

This commit introduces changes that:
- adds functionality to update/sunc the state of a MUCUser instance back to the cache;
- explicitly document that code that obtains a MUCUser from a cache is responsible for syncing it's modified state back in;
- make existing users fo MUCUser sync back any modified instances to the cache.