OF-2229 OF-2230: Send presence leave when cluster nodes leave
When observing another cluster node leaving the cluster, or when the local node leaves the cluster, occupants of MUC rooms that are connected to the local node will no longer be able to communicate with occupants of the same chat room that are connected to the now disconnected cluster node(s). For these occupants, presence 'leave' stanzas should be sent to the local occupants, so that their clients properly display that these occupants have now 'left'.