OF-2224 OF-2227 OF-2228 OF-2229 OF-2230 OF-2231 OF-2233
First stab at reconciliation of ROOM_CACHE and the local equivalents on each node, when cluster nodes join or leave a cluster.

Additionally, this commit should send out appropriate 'join' and 'leave' presence stanzas for occupants of rooms that are now, or no longer, in the room as a result of their cluster node joining or leaving the cluster.