OF-1840: Change log file usage
This drops the separation of logged message over different log files by log level. Effectively, this removes all but the 'all.log' file, renaming that to Openfire log.

A drop-down selection has been added, to retain functionality for those that desire to look at only a singular severity of logged messages.

Additionally, this commit adds ANSI-based color to the log level as printed on each line. This provides a very helpful visual hint, drawing attention to the more important log lines.

This change introduces ANSI control characters to the log file. Native readers often can handle these (for example, 'less' has the -R option for this). It's good to remember that the log configuration can be easily modified by end users, by modifying the log4j2.xml configuration file with their own preferences (which can be used to undo this change).

Corresponding changes have been made to the admin console. As the default ANSI coloring as applied by Log4j2 works best on a dark background, the background of the displayed log is changed to be dark.