Support caches for which entries are not directly relatable to exactly one cluster node.
A lot of caches hold information that relate to exactly one cluster node (eg: a client session that is physically connected to a server). This is not always the case. The Components cache, for example, tracks component connections (either external or internal). Components of the same name are allowed to connect to multiple cluster nodes.

The entry listener that we've been using so far assumes the former scenario, and breaks on the latter. This commit introduces a potential fix, that is based on a Function that takes the new entry value, and determines based on that to which cluster nodes the entry now relates.