OF-2281: Prevent invalid sublist
At times, a negative value is provided as the 'max amount of items' (probably signifying that there's no limit), when asking for a collection of published pubsub items for a particular node. In that case, an exception occurred while trying to get the sub-collection of the total collection of items to be returned.