Clean up 'local' stored MUC data when room is destroyed
When a room is destroyed, then the cluster node on which this occurs will clean up the cache, as well as its locally stored data. Any locally stored data on other cluster nodes should also be cleaned up.

This commit introduces a cache entry event listener that performs this clean up when it detects that a cache entry is removed.