OF-2297: Gracefully remove 'lost' MUC Rooms (clustering)
When a cluster node breaks out of the cluster, it sometimes occur that it locally does not have access to all cache content. A common way to resolve issues surrounding this is to maintain a local copy of relevant data.

For MUC rooms, this approach is not feasible, as it would require a continuous synchronization of MUC room data (which is resource intensive).

Instead of 'recovering' lost cache data, this commit considers a MUC room that has disappeared from the cache (during a cluster leave event) to be 'broken'. Presence unavailable (with status code 333) is sent to all local occupants, indicating that they're no longer in the room.