OF-2298: Add status code 333 when MUC occupant leaves caused by cluster breakage
When a cluster breaks, MUC occupants that are connected to now unavailable cluster nodes should be considered as having 'left' the room. This commit adds a status code to the corresponding presence update that indicates that this departure is due to a technical issue. See https://xmpp.org/extensions/xep-0045.html#service-error-kick