OF-2300: Detect and recover 'lost' cache data after cluster breakage
When a cluster node drops out of a cluster, it might be unable to retrieve cache entries. This means that those cache entries are effectively lost.

To work around this, relevant caches are supported by local data structures that can be used to re-populate these caches.

This commit detects and recovers entries that is lost in caches (as described above) for all caches in SessionManager and RoutingTableImpl