OF-2311: Remove resumable state when closing remote sessions
When a cluster node closes a (client) session that lives on a different cluster node, it's is very likely that this session is never going to be resumed (XEP-0198): the other node presumably has a good reason to want to close it (eg: replacing it with another session in IQBindHandler).

This commit ensures that a session on a different cluster node does not fall in a 'resumable' state when it is closed by another cluster node.