OF-2313: Do not depend on clustered cache when resolving cluster outage
Code that facilitates working with incoming server sessions previously depended on two caches. The 'second' cache stored 'validatedDomains' information. This comment was added:

> Content is stored in a clustered cache so that even in the case of the node hosting the sessions is lost we can still have access to this info to be able to perform proper clean up logic

We now know that in case of an incident that involves losing a cluster node, the content of caches cannot be depended on (instead, local copies of cached data needs to be maintained).

This commit removes the second cache, replacing it with a construct that keeps 'validatedDomains' info in both the primary cache, as well as locally-stored data.