OF-2340: send stanzas that are 'responses' through PacketRouter, not RoutingTable
In various places in the code, Openfire generates a response to a stanza, to be delivered back to the sender of the original stanza. Typically, this involves errors, such as delivery errors.

Some of the existing code uses the RoutingTable to deliver these stanzas. This works, but bypasses a lot of checks and flags (as well as the interceptors). For "new" stanzas, it's more appropriate to use PacketRouter, as this will more closely resemble the normal stanza processing.