OF-2353: Disable lookups in log4j2 message converter
This adds the {nolookups} directive to the log4j message directive, which disables lookups on another level (CVE-2021-44228).

Note that after upgrading to log4j2 2.16.0, this directive will be ignored by log4j2, as the lookup functionality has been removed completely in that version. A message might be logged to that effect when starting up.