OF-2358: When room is destroyed, consider that room might be gone
In MUC code that deals with room destruction, the 'room' instance is considered 'nullable' (in other words: it might not be there anymore).

The same code has a fall-back routine that intends to deal with cluster-related events. Prior to this commit, this unconditionally assumes that a local copy of the room will be present.

The former seems to contradictd the latter. This commit adds a null-check to the latter.