for OF-2367: Log exception when manual session close fails
When an admin attempts to close a session and this fails, the resulting exception should be logged.

Ignoring exceptions assumes that a certain condition is present, which might not be the case. Without the exception being logged, there is no way to find out. Exceptions should always be logged - even if they're expected.