for OF-2367: Do not retain reference to Session for a long time
When a client session has finished resource binding, Openfire will send a version query after some delay.

This commit prevents a reference to the Session object to be retained while waiting for that delay to pass. Instead, the session address is stored, and used to re-obtain the session after the delay has expired.

There is (unproven) concern that retaining a reference to the session instance is a factor of OF-2367. With this commit, this no longer is a concern.