OF-2380: Replace List-based cache values with Set in InMemoryPubsubPersistenceProvider
The List-based values make for inefficient lookups, while the order that they define is not used (when OF-2379's fix is merged).

This commit replaces the List with a HashSet, which improves performance.

Note that this commit introduces a change to the Cache definition used by clustering. This shouldn't be included in a minor/patch Openfire release.