OF-2381: Fixes for MUC 'ghost user' and 'idle user' detection
A 'ghost' occupant (one that is no longer reachable) should be detected by an IQ ping (or routing issues), which is now tied to the 'Check if users are still connected after they have been idle for X minutes' in the admin console (defaults to 'enabled', 8 minutes)

Removing an 'inactive' user (in context of the 'Kick users after they have been idle for Y minutes' admin console setting) should _not_ take into account Ghost detection, as the original intent of this feature appears to relate to connected, but otherwise inactive users.

Kicking a 'ghost' that fails to respond to an IQ Ping should be tied to the IQ Ping request/response (and not, as previously implemented, by detecting continued inactivity for some period after a Ping request should have been sent).

Kicking an occupant doesn't require an additional broadcast. This only duplicates presence stanzas in the room.

The interval with which checks are executed should be tied to the maximum allowed idleness (to prevent cases where Openfire checks once per day for users that have been inactive for longer than one minute, for example).