Group Sharing: rework caching
Warning: this introduces a potential breaking change to proprietary sub-classes of AbstractGroupProvider that are _not_ read-only (if anyone that maintains such an implementation and reads this, get in contact and I'll buy you a beer). After this commit, these implementations should explicitly call the 'super' methods that they override, to ensure that the cache of shared groups remains consistent. The changes that are needed are similar to the changes that are applied in this commit to DefaultGroupProvider.

Maintenance of the cache in which metadata related to group sharing is stored is now managed 'directly', rather than as dependent on an event listener.

Cache maintenance as implemented in this commit is very simple: on any change, the the entire cache is cleared. This can probably be improved on, but that will increase complexity, which increases the risk of bugs. As changes to groups are expected to be infrequent, the blunt 'clear all' is expected to be a acceptable trade-off.

Additional unit tests have been added to verify the behavior of caches. Specifically, these new tests try to warm up a cache, then apply a change to state that is expected to be cached, then retrieves the entity that's expected to was cached again. The test then asserts that the change is visible.