OF-2381: Prevent high-frequency MUC ghost detection pings
When an occupant of a MUC room is inactive, Openfire can send it pings to see if it is still alive.

The existing implementation would generate a ping during each periodic check, after an occupant was inactive, which is to frequent. Subsequent pings should not be sooner than the configured allowed idle time.