OF-2390: Add warning to admin console for to low ghost detection
When Openfire is configured to remove occupants as MUC ghosts for a value that is shorter than the Stream Management maximum detach time, then detached users run a considerable amount of risk to be misidentified as ghosts. This commit adds a warning to the admin console when such a configuration exists. The warning might not be fully accurate (as there's no telling what maximum detach time remote users have), but can be expected to be 'generally ok'.