OF-2390: Increase default frequency for MUC ghost detection.
The intent of this change is to increase the maximum timeout of IQ Ping requests used in MUC ghost detection. This has been hard-coded to be 25% of the frequency.

A more flexible solution would involve adding a new property that controls the IQ Ping timeout (instead of using the hardcoded 25%), but that'd take some more effort to implement. This change here will have comparable effects, without needing to worry if we'd have to expose a new property on the admin console, etc.