OF-2395: Remove deprecated code
This commit removes code that was marked as being deprecated, if it was straightforward to remove / replace.

These changes are likely to break older code, but at some point, we'll need to bite the bullet, as the accumulated technical debt was getting pretty high (I've removed code that was marked as deprecated almost a decade ago).