OF-2398: Admin console: add visual indication of MUC room being locked
MUC rooms can be 'locked', which typically occurs when they are initially being created and configured. When a MUC room is 'locked', it cannot be entered by users.

The Openfire admin console should make it possible for an administrator to see if a room is locked.

This commit replaces the value in the 'users' column on the room summary page with a 'lock' icon, and displays a warning on the room settings page, when the room is locked.