OF-2403: Memory usage should be recorded directly after GC
To prevent memory usage to include (a lot of) to-be-garbage-collected data (which leads to higher utilization numbers, which needlessly alarms some users), the reported memory usage should display memory used directly after the most recent garbage collect.

One concern that I have is that this change introduces a reference to a class in the com.sun hierarchy of packages (specifically: GarbageCollectionNotificationInfo). I do not believe that all classes in this hierarchy are considered 'closed API', but I've not been able to determine conclusively if GarbageCollectionNotificationInfo is. If it is, then its usage might cause problems when Openfire is using a non-Oracle JDK. The class is annotated with jdk.Exported, which I believe makes it safe to use.