OF-2404 / OF-38: Inbound presence 'subscribe' for preexisting contact MUST be auto-responded
RFC 6121 section 3.1.3 mandates that when processing an inbound presence subscription request, if the contact exists and the user already has a subscription to the contact's presence, then the contact's server MUST auto-reply on behalf of the contact by sending a presence stanza of type "subscribed"

Openfire currently does not do this. This behavior was removed as part of OF-38, to prevent a loop of presence stanzas that occur when an RFC3921-style client acknowledges receipt of a ‘subscribed’ presence stanza by sending a ‘subscribe’ stanza.

This commit changes the fix for OF-38 by moving the supression of the stanza to code that processes the acknowledging 'subscribe'. It was previously handled in context of the intended recipient receiving it. It now is processed in context of the originator sending it off. With that change, the code that processes subscription requests in context of the recipient can again auto-respond to a stanza.

In case this change does cause issues, the auto-replying mechanism can be disabled by setting a new property named 'xmpp.presence.suppress-subscribe-autoreply' to 'true'