Run the unit tests in isolation
Openfire's internals have many statics that are hard to 'unset'. As a result, unit tests are very likely to affect one another.

To prevent this kind of interaction, this commit configures the unit test execution to happen in isolation, using a new for for every test.

This trades execution speed for reliability.