OF-2409: Fix caseing of CSS identifier
In OF-2409, the DOCTYPE of the admin pages has changed to HTML 5. This apparently has the side-effect that CSS identifiers have become case sensitive.

The change in this commit fixes the issue of the build/login/clustering status information block to move back to the top-right corner of each page.