OF-2412: Handle secure XML property migration synchronously
When Openfire starts, all XML properties are migrated to the database. This currently is performed in an asynchronous process. This seems to open the door for a race condition. As the amount of properties involved is unlikely to be very high, switching to a synchronous process seems to offer more predictability.