LGTM: Do not synchronize on a String
Instead of a String-based mutexes (which are fairly easy to accidentally duplicate in unrelated parts of code), use a mutex that is guaranteed to be scoped to the code that should be executed under protection of a lock.

The String that was used here was 'namespaced', which makes it unlikely that a problem would have been introduced here. The replacement is safer.