OF-2426: Refactor Group implementation
Reduce complexity by centralizing the responsiblity of dispatching events to event listeners and cache updates.

Avoid using the event listener mechanism that's also being provided to users of this API in the implementation itself.

As a byproduct of the latter, this fixes OF-2426: an event dispatched by the implementation no longer re-populates the cache unexpectedly.