Rename 'principal' and 'username' to 'authcid' and 'authzid'
The existing code uses 'principal' and 'username' to distinghish between the identity whose password will be used and the identity to act as.

These terms are fairly specific to Java. This commit replaces them with 'authentication identity' (or authcid) and 'authorization identity' (authzid) respectively, which are more commonly used in context of SASL authentication.

Additional minor changes are made to log messages and javadoc. No functional changes are intended to be introduced by this commit.