OF-2443: SASL PLAIN should use AuthorizationMapper when no authzid is provided
When a client authenticates, but does not provide an identity to act as ('authorization identity'), then the server needs to determine one. Typically, this is the same value as the authentiation identity (whose password was used), but there are subtleties to this.

Openfire provides a pluggable framework for this purpose. This framework should be used by all authentiation methods, but was not used by the SASL PLAIN implementation. That caused this mechanism to behave differently from others. In a default configuration, it disallows users to log in using a bare JID (as opposed to a username/localpart), while other mechanisms allow for this.