OF-2444: Refactor HttpSession#connectionQueue to strictly contain usable connections
Prior to this change, org.jivesoftware.openfire.http.HttpSession#connectionQueue could contain connections that were already consumed.

This commit refactors this field from a list to a queue, in which only 'usable' connections are kept.

The original thought behind this change was to switch to a thread-safe data type for this field, which would reduce the need for explicit locking. This is not achieved by this commit.

Even though the original goal of the change is not achieved, the change does reduce a bit of complexity, which hopefully leads to a simpler locking strategy down the road.

While evaluating usage, various issues in the code have been identified. This commit does not fix them directly. Instead, a TODO or FIXME comment was added and a JIRA ticked was raised for each of these.