OF-2458: Guard access to field to prevent ConcurrentModificationException
The internal field 'presence' should not be modified when read, as that has been observed to lead to ConcurrentModificationException being thrown.

This commit adds a lock to all access to the 'presence' and related 'extendedInformation' fields, to prevent this from happening.

Additionaly, the instance returned by the `getPresence()` method has been replaced with a defensive copy (much of the invoking code did immediately create such a copy after acquiring the instance, which now has been removed).