OF-2449: Add configuration to disable BOSH pause range check
Using the property `xmpp.httpbind.client.maxpause.ignore-invalid`, the newly introduced behavior (kill BOSH session if it provides an invalid pause value) can be disabled. This is intended to be used only for clients that depend on the previous behavior.

Additionally, a check for negative values was added to the pause value validity check.