OF-2474: IP-based access control for the admin console
This adds an 'allow' and 'block' list of IP(v4) addresses, that controls access to the admin console.

A third property is added that can be used to configure if these lists should also be applied when accessing pages that are otherwise 'excluded' from access control by the AuthCheckFilter.

An admin console page has been added to easily configure the new properties.