OF-2467: Do not reload admin page when changing MUC affiliation dropdown
The admin console page that allows you to change MUC room permissions / affiliations has a drop-down box that lists all possible affiliations. Whenever that drop-down is changed, the page used to reload, causing all modifications to be lost. This was done to show or hide the 'nickname' input field.

This commit shows/hides that field with javascript magic, removing the need for a page reload.