OF-2482: Fix invalid description element in admin-sidebar.xml
This issue has been present for a couple of years apparently, but I'm only recently seeing the admin console pages that it affects throwing an error. Maybe some dependency update in 4.8 introduces more strictness? In any case, having this fixed is a good thing, as it should help show tooltips for the affected menu items.