OF-2486: Ensure mediated MUC invite has a 'from'.
XEP-0045 mandates a 'from' attribute on the 'invite' element, used for mediated invitations. This aught to be the address of the inviter. In some scenarios, there is not an identifiable XMPP entity that is the inviter (eg: when an invitation is generated through the admin console or REST api). In such cases, this commit will use the bare JID of the room itself as the invitation. Reasoning that it can be thought of as the inviter, and having at least some value is less likely to break clients that expect the attribute to be present.