OF-2489: Reduce log level severity of SOCKS5 error
Whenever a peer sends unexpected data to the file transfer proxy, an exception occurs. As this is not something that can be 'fixed' by an Openfire administrator, the error shouldn't be logged as verbosely as on level ERROR.

This commit reduces the log level on which this is logged, and improves the wording of the logged message a little.