OF-2490: Better handle PEP requests to non-local/non-registered users
Only local, registered users of Openfire have a PEP service. Other users do not. When a client tries to interact with such a non-existing PEP service, the current behavior of Openfire is to very verbosely log this as an (unexpected) error.

This code tries to properly handle this scenario, by returning a nice error back to the requestee, and logging it less verbosely.