OF-2498: Reduce MUCRole cache size calculations
Whenever a MUCRole instance is being added to a cache, its size is being calculated. This is rather resource intensive.

MUCRole instances are largely immutable: the cache size is unlikely to change a lot. To reduce the amount of recalculations, this change retains the last calculated size for an instance, resetting it only when a change has been applied.