OF-2499: Allow MUC history to be loaded 'on demand'
Now that MUC history is stored in a cache, this cache might unexpectedly _not_ contain data. This typically occurs around events that relate to joining or leaving a cluster.

With this change, the room history is loaded on demand from the database when it's not already present in the cache.