OF-2475: Allow 'future' users
Openfire supports mechanisms (which are unlikely to be used by default) that allow User representations to be created on an ad-hoc basis: typically when the user successfully authenticates. This leaves some room for users to be interacted with, prior to Openfire realizing that the user actually exists.

This commit introduces a new feature that allows stanzas addressed to users that could be potential future users to be handled as if that user exists, but is offline.

The feature can be enabled using the new system property `usermanager.future-users.enable`