OF-2476: Trunking/Gateway support
Openfire can be used to accept connections on behalf of remote domains, acting as a 'trunk' or 'gateway' for/to that domain. This feature can be useful when centralized orchestration of data (monitoring, filtering) is needed. This commit adds basic support and documentation for this feature.

To enable the functionality, all participating servers need to have specific configuration. This is explained in documentation that is added in this commit.