OF-2514: Differentiate between empty and missing SASL initial response
XMPP defines a difference between not having an initial response (eg: `<auth></auth>`) and having an intial response that is present, but explicitly empty (`<auth>=</auth>`).

Some SASL mechanisms should make use of this. EXTERNAL, for example, should query if the initial response is missing, but when it's empty, it should continue by using a value obtained elsewhere.

A more detailed problem definition is provided in https://igniterealtime.atlassian.net/browse/OF-2514

As Java's SaslServer API does not facilitate differentiating between the two (it does not support `null` values), this commit adds a flag to the peer's Session to indicate what the state of the initial response is.