OF-2513: Do not require authzid on SASL EXTERNAL for S2S
Openfire should not require inbound federation connections to include an authorization identity in the initial response of the SASL EXTERNAL mechanism-based authentication attempt. RFC 6120 clearly states that this is optional data.

This commit removes the requirement, but adds a configuration property (`xmpp.auth.external.server.require-authzid`) that can be used to re-enable the requirement, if so desired.

Additionally, another new property (`xmpp.auth.external.server.skip-sending-authzid`) has been added that will prevent Openfire from _sending_ an authzid in the SASL initial handshake when attempting to connect to remote domains using EXTERNAL, sending an empty initial response (`=`) instead. The default behavior is to send the authzid, as prior to this commit. This is recommended by XEP-0178. This property was added mainly to facilitate testing of the other change in this commit.