OF-2521: Allow SASL EXTERNAL for S2S by default
For SASL EXTERNAL to be used to authenticate server-to-server connections, Java's "client authentication" must be enabled, allowing peer certificates to be processed.

This improves on the changes in OF-1855, which would enable this by default _if_ the local server had a non-selfsigned certificate installed. That condition does not make a lot of sense, as the local certificate does/should not affect the validation of the certificates provided by the peer.