OF-2524: Add MUC message history in bulk
Instead of loading message and adding to MUC room history one-by-one, do this in a batch.

With OF-2499, adding a message requires a lock to be held. In clustering, the repeated acquiring and releasing of that lock adds notable delays.

This commit prevents much of this overhead by adding all history in one go, replacing the lock contention for each message, with one singular lock.