OF-2524: Add database index for MUC room log retrieval
When loading the MUC room history for a single room, a query is executed that is not optimized by the existing database indexes:

`SELECT sender, nickname, logTime, subject, body, stanza FROM ofMucConversationLog WHERE logTime>? AND roomID=? AND (nickname IS NOT NULL OR subject IS NOT NULL) ORDER BY logTime`

On Ignite's server, this takes more than 400 milliseconds to execute.

When room preloading is disabled, all rooms that are being used need to be loaded from memory. Looking at the admin console will load all rooms on that page. If that contains 25 rooms, then this query alone causes a delay of many seconds before the page is loaded.