OF-2556: Record namespace prefixes sent on root XML element
XMPP connections take the form of an XML document, starting with a 'stream' root element.

Openfire processes the direct child elements as if they were individual documents (for reasons beyond the scope of this issue). In that approach, any namespaces defined in the original root element are lost.

Generally, this isn't much of an issue, as a default namespace is assumed/used that is specific to the connection type. However, if the root element defined _other_ issues, that information is lost.

Defining additional namespaces is allowed in XML/XMPP, but isn't typically used. If it were, we would have hit problems long ago. There is one known application of this, which is Dialback (refactoring Openfire's S2S server to use MINA as part of OF-1112 is how this issue was found).

Openfire should be able to XML-parse child elements of the root 'stream' tags, if those child elements use a prefix that refers to a namespace defined on that root element.

This commit achieves that, by recording such declarations in the `Connection` representation over which they were sent. When such declarations are found, each stanza that is processed is wrapped in dummy root element, on which those definitions are placed, prioer to XML parsing. After parsing, the root element is discarded again.