OF-2557: Show TLS config on each session/connection
The cipher suite used for encryption is currently exposed (and shown on the admin console) only for S2S sessions, but should be exposed for all session types.

Similarly, lets expose the TLS version that is being used (to complement the cipher suite). This will be useful to diagnose the TLS configuration for each session type.

Implementation-wise, the TLS characteristics should be exposed by Connections, not by Session objects.

Confusingly, Openfire refers to ‘securing’ a connection, which could refer to authentication or encryption. Rename to ‘encrypt’ where appropriate.