Refactor StanzaHandler: introduce StreamErrorException
When validating a stream that is being set up, using a dedicated Exception instead of booleans will improve readability of the code, and will make it easier to extend the class.

This commit introduces a StreamErrorException and uses that when validating a new stream.

No functional changes are intended to be introduced by this commit.