OF-2566: Refactor WebSockets to reinstate Stream Management
With OF-963, the 'resume' feature of Stream Management (XEP-0198) was disabled, as it was buggy. The purpose of this commit is to reinstate this feature.

To do this, the implementation of WebSockets has been refactored to more closely resemble the implementation of the standard (socket/NIO-based) client
connection handling. It does so by introducing a `WebSocketClientConnectionHandler` and a `WebSocketClientStanzaHandler` that mimics the application of
`ClientConnectionHandler` and `ClientStanzaHandler` respectively, used for NIO/TCP sockets.

The motivation to make the WebSocket handling more similar to that of NIO/TCP sockets is two-fold:
- the latter is does not suffer from the Stream Management issue - it apparently does something better.
- maintenance of the code will be simplified, by having the same approach being re-used.