OF-2597 Adds config option for wildcards in auth-exclusions
Adds a new configuration option (SystemProperty<Boolean>) that controls whether wildcards can be used in auth-exclusions. Changes to the property value do not require a restart to become effective. The default value for this property is to disallow wildcard usage.

Wildcards should not be used when configuration auth-exclusions. They are easily abused.

It cannot be ruled out that existing versions of third-party code (such as plugins) depend on the ability to use wildcards in auth-exclusions. Wildcards should therefor remain usable - but perhaps not by default.