WIP: Fix classloading issue in Jetty 10.0.15
With Jetty 10.0.15, the admin console shows HTTP 503 error responses, and ClassLoading exceptions are logged.

It appears that in Jetty 10, unless a ClassLoader has been explicitly set on a ContextHandler, Jetty will use its own implementation. I assume (but did not check) that in Jetty 9, this was different (possibly, it used the ClassLoader of the invoking thread?). The ClassLoader that is provided by Jetty cannot find most of the libraries in the way they are provided by Openfire.

Openfire uses its own ClassLoaders. When Jetty's embedded server is instantiated to use the ClassLoader of the invoking thread as done in this commit, then the problem disappears.

Questions remain:
- Why does Jetty use their own ClassLoader (unless explicitly overridden)? This suggests that this is somehow important (which this change has undone - which might not be safe).
- Is there a better way of doing this? Maybe use a different ClassLoader instance, or somehow configure the Jetty classloader to find the classes it needs?
- Will this work with plugins?
- Do we need to do this elsewhere (eg: everywhere a Jetty Server is instantiated?)