OF-2611: LocalIncomingServerSession unit test should wait until 'done'.
Prior to this commit, the unit test for an inbound server session waited for a certain period of time, assuming that the test had run by that time.

In this commit, a structure is introduced that allows the test to explicitly flag the 'done' status. This should improve the time it takes to execute tests.