OF-2611: LocalIncomingServerSession unit test: define a session that's not authenticated as 'no connection'
Due to the nature of this test, it's possible for Openfire to keep open a session, while the local test has deduced that it cannot continue. Although the session is not 'null', it still isn't properly set-up.

This commit allows a session that is established, but NOT authenticated to pass the 'NO CONNECTION' definition.