Generate Javadoc only with Java 11
For a reason beyond me, the CI-profile of Maven causes the unit tests to fail (with instrumentation errors) when running on Java 17.

This removes the CI profile from the generic build phase.

The CI profile (and thus, generation of Javadoc and Sources) is still active on the 'deploy' phase. In this phase, only Java 11 is used.