OF-2624: Try to answer data forms in users' preferred language
When a user is requesting a data form, try to localize the text in the form according to the user's preferred language.

In this commit, the preferred language is being obtained from the session of a locally connected user. This will not work for federated users.

As a fallback, the default langague as configured in Openfire is used (as is already the case prior to this commit).